requirements.txt
amazon-braket-sdk==1.16.0
antlr4-python3-runtime==4.9.2
backoff==2.2.1
boltons==21.0.0
boto3==1.20.49
botocore==1.23.49
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
dill==0.3.6
fonttools==4.38.0
graphviz==0.20.1
ibm-cloud-sdk-core==3.16.1
ibm-platform-services==0.31.0
idna==3.4
Jinja2==3.0
jmespath==0.10.0
joblib==1.2.0
kiwisolver==1.4.4
lark-parser==0.12.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mpmath==1.2.1
mypy-extensions==0.4.3
nest-asyncio==1.5.6
networkx==2.8.8
ntlm-auth==1.5.0
numpy==1.24.1
openpulse==0.4.1
openqasm3==0.4.0
opt-einsum==3.3.0
oqpy==0.1.2
packaging==23.0
pandas==1.5.3
pbr==5.11.1
# Editable install with no version control (phiqonnect==1.0.1)
Pillow==9.4.0
ply==3.11
psutil==5.9.4
pycparser==2.21
pydantic==1.8.2
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytket==1.3
pytket-braket==0.20.0
pytket-qiskit==0.26.0
pytz==2022.7.1
qiskit==0.36.0
qwasm==1.0.1
requests==2.28.2
requests-ntlm==1.1.0
retworkx==0.12.1
rustworkx==0.12.1
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
stevedore==4.1.1
symengine==0.9.2
sympy==1.11.1
threadpoolctl==3.1.0
tweedledum==1.1.1
urllib3==1.26.14
websocket-client==1.3.3
websockets==10.4
